Begining of z management
Z management is implemented with
a slot that is called before saving
and is used to reorder widgets.

The problem is that this slot is only
called when position of widgets have
changed. If they have been dragged
but did not move on the grid, the
slot is currently not called.